+++ src/cli.ts
@@ -1,5 +1,5 @@
 import { program } from "commander";
-import { analyzeStatements } from "./analyzers/analyzeStatements";
+import { banki } from "./banki";
 
 export interface CLIOptions {
   dir: string;
@@ -12,7 +12,7 @@ program
   )
   .option("-s, --show-transactions, Show transactions for each category")
   .action((options) => {
-    void analyzeStatements(options);
+    void banki(options);
   });
 
 program.parse(process.argv);
