--- src/types/CCTransaction.ts
@@ -1,1 +1,0 @@
-export type CCTransaction = Record<string, string>;
