--- src/types/RecurringTransaction.ts
@@ -1,9 +1,0 @@
-export interface RecurringTransaction {
-  description: string;
-  estimatedAmount: number;
-  estimatedDayNumber: number;
-  label: string;
-  isPaid: boolean;
-  actualAmount: number;
-  actualDate: string;
-}
